platform/pop-render/service/requirements.txt
# ASO Render Service Dependencies
# Flask web framework and WSGI server
Flask==3.0.0
Gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# Object Storage
boto3==1.34.21

# Queue and Cache
redis==5.0.1
rq==1.16.1

# Image Processing
Pillow==10.2.0
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
opencv-python==4.9.0.80

# Monitoring
prometheus-client==0.19.0
